\name{corpmetrics-package}
\alias{corpmetrics-package}
\docType{package}
\title{
Corporate Finance Metrics and Basic Tools
}
\description{
This package provides essential financial functions tailored for Economics undergraduates, MBA students, and investors, enabling them to evaluate and screen investment opportunities effectively. It covers a wide range of corporate finance topics, including Key Performance Indicators (KPIs) from balance sheets and income statements, investment analysis methods such as Net Present Value (NPV) and Internal Rate of Return (IRR), valuation techniques for discounted dividends and fixed income securities. Additionally, it includes loan amortization schedules and the Capital Asset Pricing Model (CAPM).
}
\details{
\tabular{lll}{
Package: \tab corpmetrics\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2024-09-08\cr
License: \tab GPL-2\cr
}
}
\author{
Pavlos Pantatosakis \email{pantatosakisp@yahoo.com}.
}

\section{Maintainers}{
Pavlos Patantosakis \email{pantatosakisp@yahoo.com}.
}

\references{
The Investopedia Team (2024). Useful Balance Sheet Metrics. https://www.investopedia.com/financial-edge/1012/useful-balance-sheet-metrics.aspx)

Hayes A. (2024). Profitability Ratios: What They Are, Common Types, and How Businesses Use Them. https://www.investopedia.com/terms/p/profitabilityratios.asp

Picardo E. (2024). P/E Ratio vs. EPS vs. Earnings Yield: What's the Difference? https://www.investopedia.com/articles/investing/120513/comparing-pe-eps-and-earnings-yield.asp

Berk, J. B. and DeMarzo, P. M. (2017). Corporate finance. Pearson Education. - ISBN: 1292160160

Ehrhardt, M. C. and Brigham, E. F. (2010). Corporate finance: A focused approach. South-Western Cengage Learning. - ISBN: 1439078084

Jordan, B. D., Ross, S. A., and Westerfield, R. W. (2010). Fundamentals of corporate finance. McGraw Hill. - ISBN: 9780073382395

Will Kenton (2024). Capital Asset Pricing Model (CAPM): Definition, Formula, and Assumptions. https://www.investopedia.com/terms/c/capm.asp
}



